Switch int type.

+++ src/codr7/jx/libs/core/types/IntType.java
@@ -1,9 +1,7 @@
 package codr7.jx.libs.core.types;
 
 import codr7.jx.BaseType;
-import codr7.jx.IValue;
-import codr7.jx.VM;
 
-public class IntType extends BaseType<Integer> {
+public class IntType extends BaseType<Long> {
     public IntType(final String id) { super(id); }
 }
